Extra method for returning BinContent arrays.
